[streaming] Message bundle use inplacement instance (#6606)
* streaming message bundle use inplacement instance

* fix typo & enable common test

* fix compiler warning

* block copy for serilization

* add reference

* remove streaming common test to travis script

+++ streaming/src/test/mock_transfer_tests.cc
@@ -105,9 +105,9 @@ TEST_F(StreamingTransferTest, exchange_consumed_test) {
   std::shared_ptr<uint8_t> data(new uint8_t[data_size]);
   auto func = [data, data_size](int index) { std::fill_n(data.get(), data_size, index); };
 
-  int num = 10000;
+  size_t num = 10000;
   std::thread write_thread([this, data, data_size, &func, num]() {
-    for (uint32_t i = 0; i < num; ++i) {
+    for (size_t i = 0; i < num; ++i) {
       func(i);
       writer->WriteMessageToBufferRing(queue_vec[0], data.get(), data_size);
     }
